refactor: refatora código para utilizar classes de erro
