+++ src/errors/ErrorMessages.ts
@@ -4,6 +4,7 @@ export enum ErrorMessages {
   expiredToken = 'Token expirado',
   tokenNotProvided = 'Token não fornecido',
   unauthorized = 'Não autorizado',
+  emailNotRegistered = 'Email não cadastrado',
   wrongEmail = 'Email incorreto',
   wrongPassword = 'Senha incorreta',
   noPermissionForAction = 'Acesso negado. Você não tem permissão para essa ação',
